% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_free.R
\name{create_sfn}
\alias{create_sfn}
\title{Build a scale-free network}
\usage{
create_sfn(S, m)
}
\arguments{
\item{S}{The size of the network (number of components)}

\item{m}{The number of vertices that newly added components will have}
}
\value{
A scale-free network represented by a square matrix
}
\description{
Builds a scale-free network using the method of Albert and Barab\'{a}si
(Reviews of Modern Physics, Vo. 74, 2002). A saturated network of size 'm' is
first seeded, with new components being added with 'm' edges sequentially
until a network of size S is produced.
}
\examples{
eg_swn <- create_sfn(S = 8, m = 2);
}
