% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_world.R
\name{create_swn}
\alias{create_swn}
\title{Build a small world network}
\usage{
create_swn(S = 100, K = 20, beta = 0.05)
}
\arguments{
\item{S}{The size of the network (number of components)}

\item{K}{Number of edges that each vertice in the network contains}

\item{beta}{Probability that a random interaction in a regular network is
rewired (parameter p in Watts and Strogatz 1998)}
}
\value{
A small world network represented by a square matrix
}
\description{
Builds a small world network using the method of Watts and Strogatz (Nature 
vol. 393, 1998)
}
\examples{
eg_swn <- create_swn(S = 32, K = 4);
}
