% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R
\name{Evo_rand_gen_var}
\alias{Evo_rand_gen_var}
\title{Run genetic algorithms to find stability}
\usage{
Evo_rand_gen_var(max_sp, iters, int_type = 0, rmx = 0.4, C = 1)
}
\arguments{
\item{max_sp}{The maximum system component size (S) to simulate}

\item{iters}{The number of iterations (complex systems) tried for each S}

\item{int_type}{The type of interaction, including random (0), competitive 
(1), mutualist (2), or predator-prey (3)}

\item{rmx}{The standard devation of interaction (off-diagonal) elements in a
complex system}

\item{C}{The connectedness of a complex system (i.e., the probability that an
off-diagonal element does not equal zero)}
}
\value{
A table showing frequency of stable complex systems found
}
\description{
Outer function for investigating the potential for a complex system to be
stabilised by component response rates. For a stretch of values of component
number (i.e., system size, S), the function runs some number of iterations, 
and in each iteration, if a system is not initially found to be stable with
a set of component response times, then a genetic algorithm is called in 
attempt to find a stable solution.
}
\examples{
sim <- Evo_rand_gen_var(max_sp = 2, iters = 1);
}
