% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure_1.R
\name{find_bgamma}
\alias{find_bgamma}
\title{Find a stabilised system}
\usage{
find_bgamma(S = 200, C = 0.05, Osd = 0.4, iters = 10000)
}
\arguments{
\item{S}{Size of the complex system}

\item{C}{Connectedness of complex system}

\item{Osd}{Standard deviation of interaction strength}

\item{iters}{Number of random matrices to simulate}
}
\value{
If successful, a list that includes a matrix (A0) that is unstable in 
the absence of component response rate variation and a matrix (A1) that is
is identical to A0 but includes variation in component response rate and is
stable.
}
\description{
This function finds a random matrix (M) for which the complex system it
represents is unstable before the addition of variation in component
response rate, but stable after its addition.
}
\examples{
find_bgamma(S = 200, C = 0.05, Osd = 0.4, iters = 2);
}
