% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_complexity}
\alias{get_complexity}
\title{Get the complexity of matrix}
\usage{
get_complexity(mat)
}
\arguments{
\item{mat}{The matrix to be valuated}
}
\value{
The complexity of the matrix mat
}
\description{
Returns the complexity of a complex system, defined by the number of system
components (S), connectance between components (C), and standard deviation of
interaction strengths (sigma) such that complexity equals sigma * sqrt(SC)
}
\examples{
eg_mat  <- matrix(dat = rnorm(n = 16), nrow = 4);
mat_cmp <- get_complexity(mat = eg_mat);
}
