% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_sigma_sqd}
\alias{get_sigma_sqd}
\title{Get variance of interaction strengths}
\usage{
get_sigma_sqd(mat)
}
\arguments{
\item{mat}{The matrix to be valuated}
}
\value{
The variance of off-diagonal elements of 'mat'
}
\description{
Returns the variance of interaction stengths (sigma^2).
}
\examples{
eg_mat  <- matrix(dat = rnorm(n = 16), nrow = 4);
mat_cmp <- get_sigma_sqd(mat = eg_mat);
}
