% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammas.R
\name{make_gammas}
\alias{make_gammas}
\title{Make a distribution of gamma values}
\usage{
make_gammas(nn = 10, distribution = 1, mn = 1, sdd = 1)
}
\arguments{
\item{nn}{The number of elements in the returned vector}

\item{distribution}{The type of distribution to return}

\item{mn}{The mean of the vector to be returned}

\item{sdd}{The standard devation of the vector to be returned (where 
applicable)}
}
\value{
A vector with elements sampled froma set distribution
}
\description{
Makes one of five distributions of gamma values, including a constant value
(0), a uniform distribution (1), an exponential distribution (2), a beta
distribution (3), or a gamma distribution (4).
}
\examples{
gammas <- make_gammas(nn = 8);
}
