% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mat_rho}
\alias{mat_rho}
\title{Summarise the correlation of matrix elements}
\usage{
mat_rho(mat)
}
\arguments{
\item{mat}{The tot_res list output from the rand_gen_var function}
}
\value{
Correlation between matrix off-diagonal elements
}
\description{
Returns the correlation between off-diagonal elements A_{ij} and A_{ji}
}
\examples{
eg_mat  <- matrix(dat = rnorm(n = 16), nrow = 4);
sum_rand <- mat_rho(mat = eg_mat);
}
