% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R
\name{plot_evo_out}
\alias{plot_evo_out}
\title{Plot component response rate distributions}
\usage{
plot_evo_out(evo_out)
}
\arguments{
\item{evo_out}{Output of component response values}
}
\value{
A list of vectors of component response rates (gammas)
}
\description{
Plots the component response rate distributions from the largest systems that
the genetic algorithm finds to be stable. At the moment, only a six panel
plot is allowed
}
