% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure_1.R
\name{plot_Fig_1}
\alias{plot_Fig_1}
\title{Find stable bimodal systems}
\usage{
plot_Fig_1(A0, A1)
}
\arguments{
\item{A0}{Matrix (M) before addition of varying component response rate}

\item{A1}{Matrix (M) after additoin of varying component response rate}
}
\value{
A plot as in Figure 1 of the manuscript, showing distribution of 
eigenvalues before and after the addition of variation in component response
rate
}
\description{
Produces a two panel plot for a matrix (M) before and after the addition of 
varying component response rate (gamma).
}
\examples{
load(bi_pr_st);
plot_Fig_1(A0 = A0, A1 = A1);
}
