% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure_2.R
\name{plot_Fig_2}
\alias{plot_Fig_2}
\title{Plot eigenvalues of a large system}
\usage{
plot_Fig_2()
}
\value{
A plot as in Figure 2  of the manuscript, showing distribution of 
eigenvalues before and after the addition of uniform variation in component 
response rate
}
\description{
Produces a two panel plot for a matrix (M) before and after the addition of 
uniform variation in component response rate (gamma).
}
\examples{
plot_Fig_2();
}
