% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figures_3_4.R
\name{plot_stable_4}
\alias{plot_stable_4}
\title{Plot comparison of gamma = 1 versus var(gamma) (genetic algorithm)}
\usage{
plot_stable_4(dat, S_s = 40)
}
\arguments{
\item{dat}{Table of simulation results}

\item{S_s}{Maximum size of S to use}
}
\value{
A plot as in Figures 4 of the manuscript, showing the stability of
random matrices on a log scale in simulations where component response rate is
constant versus simulations in which it varies in a targetted way as a
consequence of a genetic algorithm search
}
\description{
Produces a plot comparing random matrix results
}
