% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_free.R
\name{rand_gen_sfn}
\alias{rand_gen_sfn}
\title{Find stabilised scale-free world networks}
\usage{
rand_gen_sfn(S, m_vals, iters, int_type = 0, rmx = 0.4, sigma = 0.4,
  mn = 0, dval = 1, g_dist = 1, g_mn = 1, g_sd = 1)
}
\arguments{
\item{S}{Number of components to randomise}

\item{m_vals}{The number of vertices that newly added components will have in
the scale free network (increasing this, but not S, increases the complexity
by increasing C).}

\item{iters}{Number of iterations (i.e., random matrices) per component}

\item{int_type}{Type of interaction between components including random (0),
competitor (1), mutualist (2), predator-prey (3), and cascade model (4)}

\item{rmx}{Standard deviation of population growth rates (for feasibility)}

\item{sigma}{Standard deviation of interaction strength among network elements}

\item{mn}{Mean interaction strength among network elements}

\item{dval}{Self-regulation of network elements (1 by default)}

\item{C}{Connectedness of matrices (i.e., probability of non-zero matrix 
element components.}

\item{m}{The number of vertices that newly added components will have in 
scale-free network construction}
}
\value{
A table of stability results, where rows summarise for each component
number (S) the number of stable or unstable (also, feasible and infeasible)
random matrices produced.
}
\description{
Compares scale-free networks in which variation in component response rate 
does not vary to random matrices in which this variation for 'm' to 'max_sp' 
components. This function is to be used for scale-free networks, which
are constructed using the method of Albert and Barab\'{a}si (Reviews of 
Modern Physics, Vo. 74, 2002). Off-diagonal elements have a mean of 'mn' and 
a standard deviation of 'sigma'.
}
\examples{
rand_gen_sfn(S = 24, m_vals = c(2, 4), iters = 4);
}
