% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_world.R
\name{rand_gen_swn}
\alias{rand_gen_swn}
\title{Find stabilised small world networks}
\usage{
rand_gen_swn(max_sp, iters, int_type = 0, rmx = 0.4, C = 1, by = 4,
  sigma = 0.4, Kdiv = 2, beta = 0.2, mn = 0, dval = 1, g_dist = 1,
  g_mn = 1, g_sd = 1)
}
\arguments{
\item{max_sp}{Maximum number of components to randomise}

\item{iters}{Number of iterations (i.e., random matrices) per component}

\item{int_type}{Type of interaction between components including random (0),
competitor (1), mutualist (2), predator-prey (3), and cascade model (4)}

\item{rmx}{Standard deviation of population growth rates (for feasibility)}

\item{C}{Connectedness of matrices (i.e., probability of non-zero matrix 
element components.}

\item{by}{Sequence between component number to randomise (e.g., '2': 2, 4, 6)}

\item{sigma}{Standard deviation of interaction strength among network elements}

\item{Kdiv}{Value to divide the component number by to produce the parameter
K for creating the small world network. For example, if S = 32 and K = 8, then
the small world network will be created from a regular network in which each
component is connected to 32/8 = 4 other components. This needs to be used
cautiously to avoid generating non-even values of K.}

\item{beta}{Probability that a random interaction in a regular network is
rewired (parameter p in Watts and Strogatz 1998)}

\item{mn}{Mean interaction strength among network elements}

\item{dval}{Self-regulation of network elements (1 by default)}
}
\value{
A table of stability results, where rows summarise for each component
number (S) the number of stable or unstable (also, feasible and infeasible)
random matrices produced.
}
\description{
Compares small world networks in which variation in component response rate 
does not vary to random matrices in which this variation for 2 to max_sp 
components. This function is to be used for small world networks, which
are constructed using the method of Watts and Strogatz (Nature vol. 393,
1998). Off-diagonal elements have a mean of 'mn' and a standard deviation of
'sigma'.
}
\examples{
rand_gen_swn(max_sp = 16, iters = 4);
}
