% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mats.R
\name{rand_gen_var}
\alias{rand_gen_var}
\title{Find stabilised random systems}
\usage{
rand_gen_var(max_sp, iters, int_type = 0, rmx = 0.4, C = 1, by = 1,
  sigma = 0.4, mn = 0, dval = 1, g_dist = 1, g_mn = 1, g_sd = 1)
}
\arguments{
\item{max_sp}{Maximum number of components to randomise}

\item{iters}{Number of iterations (i.e., random matrices) per component}

\item{int_type}{Type of interaction between components including random (0),
competitor (1), mutualist (2), predator-prey (3), and cascade model (4)}

\item{rmx}{Standard deviation of population growth rates (for feasibility)}

\item{C}{Connectedness of matrices (i.e., probability of non-zero matrix 
element components.}

\item{by}{Sequence between component number to randomise (e.g., '2': 2, 4, 6)}

\item{sigma}{Standard deviation of interaction strength among network elements}

\item{mn}{Mean interaction strength among network elements}

\item{dval}{Self-regulation of network elements (1 by default)}

\item{g_dist}{Gamma distribution to be used (1 by default)}

\item{g_mn}{Mean value of gamma distribution (only for g_dist 0)}

\item{g_sd}{Standard deviation of gamma distribution (only for g_dist 2-4)}
}
\value{
A table of stability results, where rows summarise for each component
number (S) the number of stable or unstable (also, feasible and infeasible)
random matrices produced.
}
\description{
Compares random matrices in which variation in component response rate does
not vary to random matrices in which this variation for 2 to max_sp 
components. This function is to be used for random networks, in which
off-diagonal elements are randomly sampled from a normal distribution with
a mean of 'mn' and a standard deviation of 'sigma'.
}
\examples{
rand_gen_var(max_sp = 2, iters = 4);
}
