% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mats.R
\name{rand_rho_var}
\alias{rand_rho_var}
\title{Find a stabilised system for correlated systems}
\usage{
rand_rho_var(S, rhos, iters, int_type = 0, rmx = 0.4, C = 1,
  sigma = 0.4, mn = 0, dval = 1)
}
\arguments{
\item{S}{Number of components to randomise
#'@param rhos Vector of correlations to simulate}

\item{iters}{Number of iterations (i.e., random matrices) per component}

\item{rmx}{Standard deviation of population growth rates (for feasibility)}

\item{C}{Connectedness of matrices (i.e., probability of non-zero matrix 
element components.}

\item{sigma}{Standard deviation of interaction strength among network elements}

\item{mn}{Mean interaction strength among network elements}

\item{dval}{Self-regulation of network elements (1 by default)}
}
\value{
A table of stability results, where rows summarise for each 
correlation (rho) the number of stable or unstable (also, feasible and 
infeasible) random matrices produced.
}
\description{
Compares random matrices in which variation in component response rate does
not vary to random matrices in which this variation for a fixed number of S
components. This function is to be used for random networks, in which
off-diagonal elements are randomly sampled from a normal distribution with
a mean of 'mn', a standard deviation of 'sigma', and a pre-specified
set of correlation structures.
}
\examples{
rand_rho_var(S = 10, rhos = c(-0.2, 0, 0.2), iters = 4);
}
