% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammas.R
\name{show_gammas}
\alias{show_gammas}
\title{Produce figure of gamma distributions}
\usage{
show_gammas(nn = 1e+06, sdd = 1, mn = 10)
}
\arguments{
\item{nn}{The number of samples from distributions to make the histogram}

\item{sdd}{The standard devation of the histograms to be made}

\item{mn}{The mean of the histograms to be made}
}
\value{
A four panel plot showing histograms of distributions
}
\description{
Produces a four panel histogram of gamma distributions from the make_gammas
funciton, including a uniform distribution, exponential distribution,
beta distribution, and gamma distribution
}
\examples{
show_gammas(nn = 1000);
}
