% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mats.R
\name{species_interactions}
\alias{species_interactions}
\title{Restrict component interaction types}
\usage{
species_interactions(mat, type = 0)
}
\arguments{
\item{mat}{The matrix to be restricted}

\item{type}{The type of restriction being made}
}
\value{
The matrix is returned with appropriately restricted interaction types
}
\description{
Restricts the interaction types of a matrix to be competitive (1), mutualist
(2), or predator-prey (3).
}
\examples{
eg_mat  <- matrix(data = rnorm(n = 16), nrow = 4);
new_mat <- species_interactions(mat = eg_mat, type = 3);
}
