% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure_1.R
\name{stab_bgamma}
\alias{stab_bgamma}
\title{Find stable bimodal systems}
\usage{
stab_bgamma(S = 200, C = 0.05, Osd = 0.4, iters = 10000)
}
\arguments{
\item{S}{Size of the complex system}

\item{C}{Connectedness of complex system}

\item{Osd}{Standard deviation of interaction strength}

\item{iters}{Number of random matrices to simulate}
}
\value{
A table showing whether matrices are unstable (0) or stable (1) before
(column 1) and after (column 2) including variation in component response rate
in eigenanalysis for some number (rows) of random matrices.
}
\description{
Produces random matrices and determines whether or not they are stable using 
eigenanalysis given two different component response rates of 1.95 and 0.05 
over some number of iterations
}
\examples{
dat <- stab_bgamma(iters = 4);
}
