% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga.R, R/random_mats.R
\name{summarise_randmat}
\alias{summarise_randmat}
\title{Summarise random matrix results from genetic algorithm}
\usage{
summarise_randmat(tot_res, fea_res, rho_res, cmplxty)

summarise_randmat(tot_res, fea_res, rho_res, cmplxty)
}
\arguments{
\item{tot_res}{The tot_res list output from the rand_gen_var function}

\item{fea_res}{The fea_res list output from the rand_gen_var function}

\item{rho_res}{The rho_res list output from the rand_gen_var function}

\item{tot_res}{The tot_res list output from the rand_gen_var function}

\item{fea_res}{The fea_res list output from the rand_gen_var function}

\item{cmplexty}{The cmplxty list output from the rand_gen_var function}
}
\value{
A table of stability and feasibility results.

A table of stability and feasibility results.
}
\description{
Takes the list output of the Evo_rand_gen_var function and summarises it into
a useable format.

Takes the list output of the rand_gen_var function and summarises it into a
useable format.
}
\examples{
sim_ga      <- Evo_rand_gen_var(from = 2, to = 2, iters = 1);
sum_rand_ga <- summarise_randmat_ga(sim_ga$tot_res, sim_ga$fea_res);
eg_rand  <- rand_gen_var(max_sp = 2, iters = 4);
sum_rand <- summarise_randmat(eg_rand$tot_res, eg_rand$fea_res);
}
