% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{visualise_network}
\alias{visualise_network}
\title{Visualise a network}
\usage{
visualise_network(mat)
}
\arguments{
\item{mat}{The matrix to be valuated}
}
\value{
The complexity of the matrix mat
}
\description{
Returns a plot of a network (square matrix)
}
\examples{
eg_mat  <- matrix(dat = rnorm(n = 100), nrow = 10);
mat_cmp <- visualise_network(mat = eg_mat);
}
