% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mats.R
\name{rand_gen_var}
\alias{rand_gen_var}
\title{Find a stabilised system}
\usage{
rand_gen_var(max_sp, iters, int_type = 0, rmx = 0.4, C = 1)
}
\arguments{
\item{max_sp}{Maximum number of components to randomise}

\item{iters}{Number of iterations (i.e., random matrices) per component}

\item{int_type}{Type of interaction between components (0 is random)}

\item{rmx}{Standard deviation of non-zero matrix element components}

\item{C}{Connectedness of matrices (i.e., probability of non-zero matrix 
element components.}
}
\value{
A table of stability results, where rows summarise for each component
number (S) the number of stable or unstable (also, feasible and infeasible)
random matrices produced.
}
\description{
Compares random matrices in which variation in component response rate does
not vary to random matrices in which this variation for 2 to max_sp 
components.
}
\examples{
rand_gen_var(max_sp = 2, iters = 4);
}
